Replace distutils and the old buffer protocol
Thank you very much for the contribution!

+++ gamera/gui/gaoptimizer/ResultPanel.py
@@ -77,7 +77,7 @@ class ResultPanel(wx.ScrolledWindow):
         self.SetSizer(gridSizer)
 
         width, height = gridSizer.GetMinSize()
-        self.SetScrollbars(10, 10, width/10, height/10)
+        self.SetScrollbars(10, 10, int(width/10), int(height/10))
 
         self.Layout()
